+++ core/src/main/scala/com/krux/hyperion/HyperionContext.scala
@@ -15,14 +15,18 @@ class HyperionContext(config: Config) {
   lazy val scriptUri = config.getString("hyperion.script.uri")
   lazy val logUri = config.getString("hyperion.log.uri")
 
+  //
   // Default values
+  //
   lazy val failureRerunMode = config.getString("hyperion.failure_rerun_mode")
   lazy val role = config.getString("hyperion.role")
   lazy val resourceRole = config.getString("hyperion.resource.role")
   lazy val region = config.getString("hyperion.aws.region")
   lazy val keypair = Try(config.getString("hyperion.aws.keypair")).toOption
 
+  //
   // EC2 default configuration
+  //
   lazy val ec2Region = Try(config.getString("hyperion.aws.ec2.region")).toOption.getOrElse(region)
   lazy val ec2KeyPair = Try(config.getString("hyperion.aws.ec2.keypair")).toOption.orElse(keypair)
   lazy val ec2AvailabilityZone = Try(config.getString("hyperion.aws.ec2.availability_zone")).toOption
@@ -35,7 +39,9 @@ class HyperionContext(config: Config) {
   lazy val ec2ImageId = config.getString(s"hyperion.aws.ec2.image.$ec2Region")
   lazy val ec2TerminateAfter = Try(config.getString("hyperion.aws.ec2.terminate")).toOption.map(Duration(_))
 
+  //
   // EMR default configuration
+  //
   lazy val emrRegion = Try(config.getString("hyperion.aws.emr.region")).toOption.getOrElse(region)
   lazy val emrKeyPair = Try(config.getString("hyperion.aws.emr.keypair")).toOption.orElse(keypair)
   lazy val emrAvailabilityZone = Try(config.getString("hyperion.aws.emr.availability_zone")).toOption
@@ -49,8 +55,14 @@ class HyperionContext(config: Config) {
   lazy val emrTerminateAfter = Try(config.getString("hyperion.aws.emr.terminate")).toOption.map(Duration(_))
   lazy val emrSparkVersion = config.getString("hyperion.aws.emr.spark.version")
 
+  //
   // SNS default configuration
+  //
   lazy val snsRole = Try(config.getString("hyperion.aws.sns.role")).toOption
   lazy val snsTopic = Try(config.getString("hyperion.aws.sns.topic")).toOption
 
+  //
+  // Graphviz styles
+  //
+  lazy val graphStyles = config.getConfig("hyperion.graphviz.styles")
 }