+++ core/src/main/scala/com/krux/hyperion/activity/MainClass.scala
@@ -6,6 +6,8 @@ import com.krux.hyperion.adt.HString
 
 class MainClass private (name: String) {
   override def toString = name
+  val fullName: String = name
+  val simpleName: String = name.split('.').last
 }
 
 object MainClass {
@@ -15,9 +17,9 @@ object MainClass {
   implicit def anyToMainClass(a: Any): MainClass = MainClass(a)
 
   def apply(mainClass: Any): MainClass = mainClass match {
-      case s: String => new MainClass(s.stripSuffix("$"))
-      case c: Class[_] => apply(c.getCanonicalName)
-      case mc => apply(mc.getClass)
+    case s: String => new MainClass(s.stripSuffix("$"))
+    case c: Class[_] => apply(c.getCanonicalName)
+    case mc => apply(mc.getClass)
   }
 }