+++ examples/src/main/scala/com/krux/hyperion/examples/ExampleSpark.scala
@@ -1,5 +1,6 @@
 package com.krux.hyperion.examples
 
+import scala.language.postfixOps
 import com.krux.hyperion.action.SnsAlarm
 import com.krux.hyperion.activity.{SparkJobActivity, SparkActivity, SparkStep}
 import com.krux.hyperion.common.S3Uri
@@ -34,7 +35,7 @@ object ExampleSpark extends DataPipelineDef {
 
   override def parameters: Iterable[Parameter[_]] = Seq(location, instanceType, instanceCount, instanceBid)
 
-  val dataNode = S3DataNode(s3 / "some-bucket" / "some-path/")
+  val dataNode = S3DataNode(s3 / "some-bucket" / "some-path" /)
 
   // Actions
   val mailAction = SnsAlarm()