+++ scripts/run-spark-step.sh
@@ -16,11 +16,25 @@ fi
 
 # Parse the spark options as any option until the first non-option (the jar)
 SPARK_OPTIONS=""
+SPARK_MASTER_OPTION="yarn-client"
+SPARK_DRIVER_MEMORY_OPTION="9g"
 
 while [[ $# > 0 ]]; do
   key=$1
 
   case ${key} in
+    --master)
+      shift
+      SPARK_MASTER_OPTION="$1"
+      shift
+      ;;
+
+    --driver-memory)
+      shift
+      SPARK_DRIVER_MEMORY_OPTION="$1"
+      shift
+      ;;
+
     --*)
       shift
       SPARK_OPTIONS="${SPARK_OPTIONS} ${key} $1"
@@ -53,4 +67,4 @@ LOCAL_JAR="${LOCAL_JAR_DIR}/${JAR_NAME}"
 # Download JAR file from S3 to local
 hadoop fs -get ${REMOTE_JAR_LOCATION} ${LOCAL_JAR}
 
-exec ${EMR_SPARK_HOME}/bin/spark-submit --master yarn-client --driver-memory 9g --class ${JOB_CLASS} ${SPARK_OPTIONS} ${LOCAL_JAR} $@
+exec ${EMR_SPARK_HOME}/bin/spark-submit --master ${SPARK_MASTER_OPTION} --driver-memory ${SPARK_DRIVER_MEMORY_OPTION} ${SPARK_OPTIONS} --class ${JOB_CLASS} ${LOCAL_JAR} $@